test: move pr468, pr798, and pr476_parse_* tests to test/frontend
Refs #1152

Made-with: Cursor
