+++ test/frontend/pr476_parse_enum_helpers.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest';
 
-import type { Diagnostic } from '../src/diagnosticTypes.js';
-import { parseEnumDecl } from '../src/frontend/parseEnum.js';
-import { makeSourceFile, span } from '../src/frontend/source.js';
-import { parseProgram } from '../src/frontend/parser.js';
+import type { Diagnostic } from '../../src/diagnosticTypes.js';
+import { parseEnumDecl } from '../../src/frontend/parseEnum.js';
+import { makeSourceFile, span } from '../../src/frontend/source.js';
+import { parseProgram } from '../../src/frontend/parser.js';
 
 describe('PR476 enum parser extraction', () => {
   const file = makeSourceFile('pr476_parse_enum_helpers.zax', '');
