+++ test/frontend/pr476_parse_extern_block_helpers.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
-import type { Diagnostic } from '../src/diagnosticTypes.js';
-import { parseTopLevelExternDecl } from '../src/frontend/parseExternBlock.js';
-import { parseParamsFromText } from '../src/frontend/parseParams.js';
-import { parseProgram } from '../src/frontend/parser.js';
-import { makeSourceFile, span } from '../src/frontend/source.js';
+import type { Diagnostic } from '../../src/diagnosticTypes.js';
+import { parseTopLevelExternDecl } from '../../src/frontend/parseExternBlock.js';
+import { parseParamsFromText } from '../../src/frontend/parseParams.js';
+import { parseProgram } from '../../src/frontend/parser.js';
+import { makeSourceFile, span } from '../../src/frontend/source.js';
 
 describe('PR476 extern block parser extraction', () => {
   it('keeps extern block parsing intact', () => {
