+++ test/frontend/pr798_addr_expr_parser.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { parseProgram } from '../src/frontend/parser.js';
+import { parseProgram } from '../../src/frontend/parser.js';
 
 const parse = (text: string) => {
   const diagnostics: Array<{ message: string }> = [];
